Modify the Seller part of creating new socket
Seller creates one DatagramSocket for each Buyer's IP/port.

+++ src/mobile/app_for_test/BuyerConfig.java
@@ -8,5 +8,5 @@ public class BuyerConfig {
     public static final String DEFAULT_VPN_CLIENT_ADDR  = "10.13.10.2";
     public static final String DEFAULT_VPN_GATEWAY_ADDR = "10.13.10.1"; // Do we need this?
     public static final int	   DEFAULT_LOCAL_PORT		= 9030;
-    public static final long   DEFAULT_UDP_TIMEOUT		= 5000;
+    public static final int    DEFAULT_UDP_TIMEOUT		= 5000;
 }
